“管理书页面界面”添加domainId字段,查询条件允许一个或多个

+++ src/main/java/com/demofactory/syscontrol/controller/admin/management/BookController.java
@@ -3,12 +3,11 @@ package com.demofactory.syscontrol.controller.admin.management;
 import com.demofactory.syscontrol.api.BookService;
 import com.demofactory.syscontrol.domain.Books;
 import org.apache.dubbo.config.annotation.Reference;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.apache.commons.lang.StringUtils;
 
 import java.util.List;
+import java.util.Objects;
 
 /**
  * @author : Hanamaru
@@ -29,18 +28,17 @@ public class BookController {
             System.out.println("result------书名不能为空");
             return "书名不为空";
         }
-
+        if(Objects.isNull(books.getDomainId()))
+        {
+            System.out.println("result------域不能为空");
+            return "域不能为空";
+        }
         return bookService.insertBook(books);
     }
 
     @PostMapping("selectBook")
     public List<Books> selectBook(Books books)
     {
-        if(StringUtils.isBlank(books.getBookName()))
-        {
-            System.out.println("result------书名不能为空");
-            return null;
-        }
         return bookService.selectBook(books);
     }
     @PostMapping("deleteBook")
@@ -50,6 +48,11 @@ public class BookController {
         {
             System.out.println("result------书名不能为空");
             return "书名不能为空";
+        }
+        if(Objects.isNull(books.getDomainId()))
+        {
+            System.out.println("result------域不能为空");
+            return null;
         }
         return bookService.deleteBook(books);
     }
